Resolution des erreurs et des fuites de memoire jusqu'a mygitAdd (ordre du main)

+++ main.c
@@ -16,43 +16,58 @@
 
 int main(int argc, char* argv[]) {
 
-	if (strcmp(argv [1], "init") == 0) {
+	if (argc > 1 && strcmp(argv[1], "init") == 0) {
  		initRefs();
  		initBranch();
 	}
 
-	if (strcmp(argv[1], "refs-list") == 0) {
- 		printf( "REFS : \n" );
+	if (argc > 1 && strcmp(argv[1], "refs-list") == 0) {
 		if (file_exists(".refs")) {
 			List* L = listdir(".refs");
+			printf("REFS :\n");
 			for (Cell* ptr = *L; ptr != NULL; ptr = ptr -> next) {
 				if ( ptr -> data[0] == '.' )
 					continue;
-				char* content = getRef( ptr -> data );
-				printf ( "− %s \t %s \n" , ptr -> data , content );
+				char* content = getRef(ptr -> data);
+				printf("−%s\t%s\n" , ptr -> data, content);
+				free(content);
 			}
+			freeList(L);
+		} else {
+			printf("Il n'y a pas de reference.\n");
 		}
 	}
 
-	if ( strcmp(argv[1] ,"create-refs") == 0) {
-		createUpdateRef(argv[2], argv[3]);
+	if (argc > 1 && strcmp(argv[1], "create-refs") == 0) {
+		if (argc > 3)
+			createUpdateRef(argv[2], argv[3]);
+		else
+			printf("Veuillez renseigner le nom de la reference et son hash.\n");
 	}
 	
-	if ( strcmp(argv[1] , "delete−ref") == 0) {
-		deleteRef(argv[2]);
+	if (argc > 1 && strcmp(argv[1], "delete-ref") == 0) {
+		if (argc > 2)
+			deleteRef(argv[2]);
+		else
+			printf("Veuillez renseigner le nom de la reference.\n");
 	}
 
-	if ( strcmp(argv[1], "add") == 0) {
-		for (int i = 2; i < argc; i ++) {
-			myGitAdd(argv[i]);
+	if (argc > 1 &&  strcmp(argv[1], "add") == 0) {
+		if (argc > 2) {
+			for (int i = 2; i < argc; i ++) {
+				myGitAdd(argv[i]);
+			}
+		}
+		else {
+			printf("Veuillez renseigner le(s) fichier(s) a ajouter.\n");
 		}
 	}
 
-	if ( strcmp(argv[1], "clear-add") == 0) {
+	if (argc > 1 &&  strcmp(argv[1], "clear-add") == 0) {
 		system("rm .add");
 	}
 
-	if ( strcmp(argv [2], "add-list") ==0) {
+	if (argc > 1 && strcmp(argv[1], "add-list") ==0) {
 		printf("Zone de preparation : \n");
 		if (file_exists(".add")) {
 			WorkTree* wt = ftwt(".add");
@@ -60,45 +75,60 @@ int main(int argc, char* argv[]) {
 		}
 	}
 
-	if (strcmp(argv[1], "commit") == 0) {
-		if (strcmp(argv[3], "-m") == 0) {
-			myGitCommit(argv[2], argv[4]);
-		} else {
-			myGitCommit(argv[2], NULL);
-		}
+	if (argc > 1 && strcmp(argv[1], "commit") == 0) {
+			if (argc > 4 && strcmp(argv[3], "-m") == 0) {
+				myGitCommit(argv[2], argv[4]);
+			} else if (argc > 2) {
+				myGitCommit(argv[2], NULL);
+			} else {
+			printf("Veuillez renseigner le nom du commit [et le message].\n");
+			}
 	}
 
-	if (strcmp(argv[1], "get-current-branch") == 0) {
+	if (argc > 1 && strcmp(argv[1], "get-current-branch") == 0) {
 		printf ("%s", getCurrentBranch());
 	}
 
-	if (strcmp(argv[1], "branch") == 0) {
-		if (!branchExists(argv[2])) {
-			createBranch(argv[2]);
+	if (argc > 1 && strcmp(argv[1], "branch") == 0) {
+		if (argc >2) {
+			if (!branchExists(argv[2])) {
+				createBranch(argv[2]);
+			}
+			else {
+				printf("La branche existe deja.\n");
+			}
+		}
+	}
+
+	if (argc > 1 && strcmp(argv[1], "branch-print") == 0) {
+		if (argc > 0) {
+			if (!branchExists(argv[2]) == 0) {
+				printf("La branche n'existe pas.\n");
+			}
+			else {
+				printBranch(argv[2]);
+			}
+		} else {
+			printf("Veuillez renseigner le nom de la branche.\n");
+		}
+	}
+
+	if (argc > 1 && strcmp(argv[1], "checkout-branch") == 0) {
+		if (argc > 2 && branchExists(argv[2]) == 1) {
+			myGitCheckoutBranch(argv[2]);
 		}
 		else {
-			printf("La branche existe deja.\n");
+			printf("Veuillez renseigner le nom d'une branche existante.\n");
 		}
 	}
 
-	if (strcmp(argv[1], "branch-print") == 0) {
-		if (!branchExists(argv[2]) == 0) {
-			printf("La branche n'existe pas.\n");
+	if (argc > 1 && strcmp(argv[1], "checkout-commit") == 0) {
+		if (argc > 2) {
+			myGitCheckoutCommit(argv[2]);
 		}
 		else {
-			printBranch(argv[2]);
+			printf("Veuillez renseigner le nom du commit.\n");
 		}
-	}
-
-	if (strcmp(argv[1], "checkout-branch") == 0) {
-		printf("La branche n'existe pas.\n");
-	}
-	else {
-		myGitCheckoutBranch(argv[2]);
-	}
-
-	if (strcmp(argv[1], "checkout-commit") == 0) {
-		myGitCheckoutCommit(argv[2]);
 	}
 
 	return 0;
